requirements.txt
appnope==0.1.2
asgiref==3.3.4
astroid==2.5.3
backcall==0.2.0
decorator==5.0.7
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.7.0
django-extensions==3.1.2
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
ipython==7.22.0
ipython-genutils==0.2.0
isort==5.8.0
jedi==0.18.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess==0.7.0
Pygments==2.8.1
pylint==2.7.4
pylint-django==2.4.3
pylint-plugin-utils==0.6
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
wcwidth==0.2.5
whitenoise==5.2.0
wrapt==1.12.1
